# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOVA Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  fullName: string;
  phone: string;
  role: UserRole;
  createdAt: string;
}

export interface RegisterRequest {
  fullName: string;
  phone: string;
  password: string;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface Category {
  id: number;
  name: string;
  nameAr: string;
  slug: string;
  icon?: string;
  image?: string;
  productCount?: number;
}

export interface CreateCategoryRequest {
  name: string;
  nameAr: string;
  slug: string;
  icon?: string;
  image?: string;
}

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  originalPrice?: number;
  images: string[];
  categoryId?: number;
  categoryName?: string;
  categoryNameAr?: string;
  stock: number;
  featured: boolean;
  badge?: string;
  createdAt: string;
}

export interface CreateProductRequest {
  name: string;
  nameAr: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  originalPrice?: number;
  images: string[];
  categoryId?: number;
  stock: number;
  featured?: boolean;
  badge?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  productId: number;
  productName: string;
  productNameAr: string;
  productImage?: string;
  fullName: string;
  phone: string;
  altPhone?: string;
  governorate: string;
  address: string;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  facebookPage?: string;
  notes?: string;
  createdAt: string;
}

export interface CreateOrderRequest {
  productId: number;
  fullName: string;
  phone: string;
  altPhone?: string;
  governorate: string;
  address: string;
  quantity: number;
  facebookPage?: string;
  notes?: string;
}

export type UpdateOrderStatusRequestStatus =
  (typeof UpdateOrderStatusRequestStatus)[keyof typeof UpdateOrderStatusRequestStatus];

export const UpdateOrderStatusRequestStatus = {
  pending: "pending",
  confirmed: "confirmed",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusRequest {
  status: UpdateOrderStatusRequestStatus;
}

export type GetProductsParams = {
  categoryId?: number;
  search?: string;
  featured?: boolean;
};