# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOVA Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderRequest {
  productId: number;
  fullName: string;
  phone: string;
  altPhone?: string;
  governorate: string;
  address: string;
  quantity: number;
  facebookPage?: string;
  notes?: string;
}