# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOVA Store API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  nameAr: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  originalPrice?: number;
  images: string[];
  categoryId?: number;
  stock: number;
  featured?: boolean;
  badge?: string;
}