# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOVA Store API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  productId: number;
  productName: string;
  productNameAr: string;
  productImage?: string;
  fullName: string;
  phone: string;
  altPhone?: string;
  governorate: string;
  address: string;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  facebookPage?: string;
  notes?: string;
  createdAt: string;
}