# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NOVA Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  originalPrice?: number;
  images: string[];
  categoryId?: number;
  categoryName?: string;
  categoryNameAr?: string;
  stock: number;
  featured: boolean;
  badge?: string;
  createdAt: string;
}